exercicios 1 ao 18

+++ sequencial/exercício_8.py
@@ -1,4 +1,4 @@
-horas_trabalhadas = float(input("Digite o numero de hosras trabalhadas:"))
+horas_trabalhadas = int(input("Digite o numero de hosras trabalhadas:"))
 valor_horas = float(input("Digite o valor da hora trabalhada:"))
 
 salario = horas_trabalhadas * valor_horas
